Small Update v08022020 - Version 0.1.1
We've changed a bit in the README.md file.
Now there's a Installation Guide, but still this project isn't finished.

Installing it now, could lead to nothing.

We've also added a .gitignore and .gitattributes.
The .gitignore is temprary, because we have not finished our work on this folder yet. The .gitattributes files is there, because the auto language detection detected the wrong programming languages for this project.

Regards!